requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bert-score==0.3.12
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
datasets==2.9.0
dill==0.3.6
editdistance==0.6.2
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
joblib==1.2.0
jsonargparse==3.13.1
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
multidict==6.0.4
multiprocess==0.70.14
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
portalocker==2.7.0
protobuf==3.20.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.7.1
PyYAML==6.0
rank-bm25==0.2.2.post2
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.2
torch==1.13.1
torchmetrics==0.8.2
tqdm==4.64.1
transformers==4.26.1
typing-extensions==4.5.0
unbabel-comet==1.1.3
urllib3==1.26.14
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
zipp==3.14.0
